[refactor] Cleaned up the Gaussian sampler method. Refactored sampling methods to own class implementing common sampler interfaces.
