+++ src/main/java/com/medals/libsdatagenerator/service/CompositionalVariations.java
@@ -2,16 +2,11 @@ package com.medals.libsdatagenerator.service;
 
 import com.medals.libsdatagenerator.controller.LIBSDataGenConstants;
 import com.medals.libsdatagenerator.model.Element;
-import com.medals.libsdatagenerator.model.SeriesStatistics;
+import com.medals.libsdatagenerator.sampler.DirichletSampler;
+import com.medals.libsdatagenerator.sampler.GaussianSampler;
 import com.medals.libsdatagenerator.util.CommonUtils;
-import org.apache.commons.rng.UniformRandomProvider;
-import org.apache.commons.rng.simple.RandomSource;
-import org.apache.commons.rng.sampling.distribution.DirichletSampler;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Random;
+import java.util.*;
 import java.util.logging.Logger;
 
 /**
@@ -20,20 +15,14 @@ import java.util.logging.Logger;
 
 public class CompositionalVariations {
 
-    public static CompositionalVariations instance = null;
-    private final CommonUtils commonUtils;
     private final Logger logger = Logger.getLogger(CompositionalVariations.class.getName());
+    private static CompositionalVariations instance = null;
+    private final CommonUtils commonUtils;
     private static final double POST_NORM_CHECK_DELTA = 0.0001;
     private static final double FINAL_SUM_TOLERANCE = 0.1;
 
-    // Enhanced services for Dirichlet sampling
-    private final MatwebDataService matwebService;
-    private final ConcentrationParameterEstimator parameterEstimator;
-
     public CompositionalVariations() {
         commonUtils = new CommonUtils();
-        matwebService = MatwebDataService.getInstance();
-        parameterEstimator = new ConcentrationParameterEstimator();
     }
 
     public static CompositionalVariations getInstance() {
@@ -43,141 +32,78 @@ public class CompositionalVariations {
         return instance;
     }
 
-    /**
-     * New Dirichlet sampling method for compositional variations
-     *
-     * @param baseComp Base composition elements
-     * @param overviewGuid GUID of the overview datasheet for series statistics
-     * @param samples Number of samples to generate
-     * @param variations List to store generated variations
-     */
-    public void dirichletSampling(List<Element> baseComp, String overviewGuid, int samples,
-                                  List<List<Element>> variations) {
+    public List<List<Element>> generateCompositionalVariations(List<Element> originalComposition,
+                                                               double varyBy, double maxDelta, int variationMode,
+                                                               int samples, String overviewGuid) {
 
-        logger.info("Starting Dirichlet sampling with overview GUID: " + overviewGuid);
-
-        // Get series statistics from overview sheet
-        SeriesStatistics seriesStats = matwebService.getSeriesStatistics(overviewGuid);
-        if (seriesStats == null) {
-            logger.severe("Failed to extract series statistics from overview sheet. Falling back to Gaussian sampling.");
-            // Fallback to Gaussian sampling
-            gaussianSampling(baseComp, 5.0, samples, variations); // Use default maxDelta of 5.0
-            return;
+        List<List<Element>> compositions = new ArrayList<>();
+        if (originalComposition == null || originalComposition.isEmpty()) {
+            logger.warning("Original composition is null or empty. Cannot generate variations.");
+            return compositions; // Return empty list, or perhaps add the (empty) originalComposition if that's desired.
         }
+        compositions.add(originalComposition); // Adding the original composition
 
-        logger.info("Successfully extracted series statistics: " + seriesStats);
-
-        // Create element order array from base composition
-        String[] elementOrder = new String[baseComp.size()];
-
-        for (int i = 0; i < baseComp.size(); i++) {
-            Element element = baseComp.get(i);
-            elementOrder[i] = element.getSymbol();
-        }
-
-        // Estimate Dirichlet concentration parameters for the base composition elements only
-        double[] concentrationParams = parameterEstimator.estimateParametersForElements(seriesStats, elementOrder);
-        if (concentrationParams == null || !parameterEstimator.validateParameters(concentrationParams)) {
-            logger.severe("Failed to estimate valid Dirichlet parameters. Falling back to Gaussian sampling.");
-            gaussianSampling(baseComp, 5.0, samples, variations);
-            return;
-        }
-
-        // Verify arrays have same length
-        if (concentrationParams.length != elementOrder.length) {
-            logger.severe("Mismatch between concentration parameters (" + concentrationParams.length +
-                    ") and element order (" + elementOrder.length + "). Falling back to Gaussian sampling.");
-            gaussianSampling(baseComp, 5.0, samples, variations);
-            return;
-        }
-
-        UniformRandomProvider rng = RandomSource.XO_RO_SHI_RO_128_PP.create();
-
-        DirichletSampler sampler = DirichletSampler.of(rng, concentrationParams);
-
-        logger.info("Created Dirichlet sampler with parameters: " + Arrays.toString(concentrationParams));
-
-
-        // Generate samples
-        int successfulSamples = 0;
-        int attempts = 0;
-        int maxAttempts = samples * 10; // Allow up to 10x attempts to account for constraint violations
-
-        while (successfulSamples < samples && attempts < maxAttempts) {
-            attempts++;
-
-            try {
-                double[] sample = sampler.sample();
-                logger.info("Sample: " + Arrays.toString(sample));
-                List<Element> variation = createElementVariation(baseComp, sample);
-
-                if (validateVariation(variation)) {
-                    variations.add(variation);
-                    successfulSamples++;
-
-                    if (successfulSamples % 100 == 0) {
-                        logger.info("Generated " + successfulSamples + " valid Dirichlet samples");
-                    }
-                } else {
-                    logger.fine("Sample failed validation, attempting another");
-                }
-
-            } catch (Exception e) {
-                logger.warning("Error generating Dirichlet sample: " + e.getMessage());
+        boolean allElementsAreFixed = true;
+        for (Element el : originalComposition) {
+            Double minComp = el.getMin();
+            Double maxComp = el.getMax();
+            if (!(minComp != null && maxComp != null && minComp.equals(maxComp))) {
+                allElementsAreFixed = false;
+                break;
             }
         }
 
-        logger.info("Dirichlet sampling completed: " + successfulSamples + " successful samples out of " +
-                attempts + " attempts");
-
-        if (successfulSamples < samples) {
-            logger.warning("Could not generate all requested samples. Generated " + successfulSamples +
-                    " out of " + samples + " requested samples.");
-        }
-    }
-
-    /**
-     * Creates an Element variation from a Dirichlet sample
-     */
-    private List<Element> createElementVariation(List<Element> baseComp, double[] sample) {
-        List<Element> variation = new ArrayList<>();
-        double totalPercentage = 0;
-        for (int i = 0; i < baseComp.size(); i++) {
-            Element baseElement = baseComp.get(i);
-            double min = baseElement.getMin();
-            double max = baseElement.getMax();
-
-            // Scale the sample to the element's allowed range
-            double newPercentage = min + (sample[i] * (max - min));
-            totalPercentage += newPercentage;
-
-            // Round to appropriate decimal places
-            double roundedValue = CommonUtils.roundToNDecimals(newPercentage, baseElement.getNumberDecimalPlaces());
-
-            Element variationElement = new Element(
-                    baseElement.getName(),
-                    baseElement.getSymbol(),
-                    roundedValue,
-                    baseElement.getMin(),
-                    baseElement.getMax(),
-                    baseElement.getAverageComposition()
-            );
-            variation.add(variationElement);
-        }
-
-        // Normalize the composition to sum to 100%
-        if (totalPercentage > 0) {
-            for (Element element : variation) {
-                element.setPercentageComposition((element.getPercentageComposition() / totalPercentage) * 100.0);
+        List<Element> effectiveComposition = originalComposition;
+        if (allElementsAreFixed) {
+            // originalComposition is not empty here due to the check at the beginning.
+            logger.info("All elements in the input composition are fixed. Applying fallback variation logic, ignoring X:X constraints for sampling.");
+            effectiveComposition = new ArrayList<>();
+            for (Element el : originalComposition) {
+                effectiveComposition.add(new Element(
+                        el.getName(),
+                        el.getSymbol(),
+                        el.getPercentageComposition(),
+                        null, // Effectively remove min/max for variation generation
+                        null,
+                        el.getAverageComposition()
+                ));
             }
         }
-        return variation;
+
+        // Generate all combinations by Uniform distribution
+        // System.out.println("\nGenerating different combinations for the input composition (refer log for list)...");
+        logger.info("\nGenerating different combinations for the input composition (refer log for list)...");
+
+
+        if (variationMode == LIBSDataGenConstants.STAT_VAR_MODE_GAUSSIAN_DIST) {
+            Map<String, Object> metadata = new HashMap<>();
+            metadata.put("maxDelta", maxDelta);
+            GaussianSampler.getInstance().sample(effectiveComposition, samples, compositions, metadata);
+
+        } else if (variationMode == LIBSDataGenConstants.STAT_VAR_MODE_DIRICHLET_DIST) {
+            Map<String, Object> metadata = new HashMap<>();
+            metadata.put("overviewGuid", overviewGuid);
+            DirichletSampler.getInstance().sample(effectiveComposition, samples, compositions, metadata);
+
+        } else { // For uniform distribution
+            // Start backtracking with an empty "current combo" and a running sum of 0
+            CompositionalVariations.getInstance().getUniformDistribution(
+                    0,
+                    effectiveComposition,
+                    varyBy,
+                    maxDelta, // 'limit'
+                    0.0,
+                    new ArrayList<Element>(),
+                    compositions);
+        }
+
+        return compositions;
     }
 
     /**
      * Validates that a variation meets all constraints
      */
-    private boolean validateVariation(List<Element> variation) {
+    public boolean validateVariation(List<Element> variation) {
         double totalPercentage = 0.0;
 
         for (Element element : variation) {
@@ -210,195 +136,7 @@ public class CompositionalVariations {
         return true;
     }
 
-    public void gaussianSampling(List<Element> baseComp, double maxDelta, int samples,
-                                 List<List<Element>> variations) {
-        Random rand = new Random();
-        int initialSize = variations.size();
-
-        while (variations.size() < initialSize + samples) {
-            double sumFixedPercentages = 0.0;
-            ArrayList<Element> tempVariableElements = new ArrayList<>();
-            // currentRawVariantElements stores elements in original order, some fixed, some are initial variable samples
-            ArrayList<Element> currentRawVariantElements = new ArrayList<>();
-
-            // First Pass: Identify fixed/variable, apply initial sampling to variables
-            for (Element baseElement : baseComp) {
-                Double minComp = baseElement.getMin();
-                Double maxComp = baseElement.getMax();
-                boolean isFixed = minComp != null && maxComp != null && minComp.doubleValue() == maxComp.doubleValue();
-
-                if (isFixed) {
-                    Element fixedElementCopy = new Element(baseElement.getName(), baseElement.getSymbol(),
-                            minComp, // Use the fixed percentage
-                            minComp, maxComp, baseElement.getAverageComposition());
-                    currentRawVariantElements.add(fixedElementCopy);
-                    sumFixedPercentages += minComp;
-                } else {
-                    double baseVal = baseElement.getPercentageComposition();
-                    // Ensure symbol exists in fallback map, otherwise skip this sample generation attempt
-                    if (!LIBSDataGenConstants.ELEMENT_STD_DEVS_FALLBACK.containsKey(baseElement.getSymbol())) {
-                        logger.warning("Symbol " + baseElement.getSymbol() + " not in ELEMENT_STD_DEVS_FALLBACK. Skipping sample generation attempt.");
-                        currentRawVariantElements.clear(); // Mark as invalid attempt
-                        break;
-                    }
-                    double stdDev = LIBSDataGenConstants.ELEMENT_STD_DEVS_FALLBACK.get(baseElement.getSymbol());
-                    double delta = rand.nextGaussian() * stdDev;
-                    delta = Math.max(-maxDelta, Math.min(delta, maxDelta));
-                    double newVal = baseVal + delta;
-                    newVal = Math.max(0, newVal); // Clamp to 0 minimum
-
-                    if (minComp != null) {
-                        newVal = Math.max(newVal, minComp);
-                    }
-                    if (maxComp != null) {
-                        newVal = Math.min(newVal, maxComp);
-                    }
-
-                    Element variableElementInstance = new Element(baseElement.getName(), baseElement.getSymbol(),
-                            newVal, minComp, maxComp, baseElement.getAverageComposition());
-                    currentRawVariantElements.add(variableElementInstance);
-                    tempVariableElements.add(variableElementInstance);
-                }
-            }
-
-            // If loop was broken (e.g. missing std dev), restart sample generation
-            if (currentRawVariantElements.size() != baseComp.size()){
-                logger.info("Skipping sample due to issue in first pass (e.g. missing std dev for a symbol).");
-                continue;
-            }
-
-            // Zero Check for Variable Elements that were non-zero initially
-            boolean variableElementBecameZero = false;
-            for (int i = 0; i < tempVariableElements.size(); i++) {
-                Element varElement = tempVariableElements.get(i);
-                // Find original base element corresponding to this tempVariableElement
-                // This is a bit indirect; assumes tempVariableElements are added in same order as variable elements appear in baseComp
-                Element originalBaseForVar = null;
-                int varCounter = 0;
-                for(Element baseElem : baseComp) {
-                    boolean isBaseFixed = baseElem.getMin() != null && baseElem.getMax() != null && baseElem.getMin().doubleValue() == baseElem.getMax().doubleValue();
-                    if (!isBaseFixed) {
-                        if (varCounter == i) {
-                            originalBaseForVar = baseElem;
-                            break;
-                        }
-                        varCounter++;
-                    }
-                }
-
-                if (varElement.getPercentageComposition() == 0 && originalBaseForVar != null && originalBaseForVar.getPercentageComposition() > 0) {
-                    logger.info("Variable Element: " + varElement.getName() + " (originally > 0%) sampled to 0. Discarding sample.");
-                    variableElementBecameZero = true;
-                    break;
-                }
-            }
-            if (variableElementBecameZero) {
-                continue;
-            }
-
-            // Normalization of Variable Elements
-            double currentSumVariableRaw = 0;
-            for (Element e : tempVariableElements) {
-                currentSumVariableRaw += e.getPercentageComposition();
-            }
-
-            double targetSumVariable = 100.0 - sumFixedPercentages;
-
-            if (targetSumVariable < 0) { // Sum of fixed elements already > 100%
-                logger.info("Sum of fixed percentages (" + sumFixedPercentages + "%) > 100%. Discarding sample.");
-                continue;
-            }
-            if (currentSumVariableRaw <= 0 && targetSumVariable > 0 && !tempVariableElements.isEmpty()) {
-                logger.info("Sum of raw variable percentages is " + currentSumVariableRaw +
-                        " but target is " + targetSumVariable + ". Cannot normalize. Discarding sample.");
-                continue;
-            }
-
-
-            double variableScalingFactor = 1.0;
-            if (currentSumVariableRaw > 0) {
-                variableScalingFactor = targetSumVariable / currentSumVariableRaw;
-            } else if (targetSumVariable == 0 && currentSumVariableRaw == 0) { // All variable elements are zero, and target is zero
-                variableScalingFactor = 0; // Effectively keeps them zero
-            }
-
-
-            for (Element varElement : tempVariableElements) {
-                double scaledVal = varElement.getPercentageComposition() * variableScalingFactor;
-
-                // Re-clamp against min/max and ensure >= 0
-                scaledVal = Math.max(0, scaledVal);
-                if (varElement.getMin() != null) {
-                    scaledVal = Math.max(scaledVal, varElement.getMin());
-                }
-                if (varElement.getMax() != null) {
-                    scaledVal = Math.min(scaledVal, varElement.getMax());
-                }
-                varElement.setPercentageComposition(CommonUtils.roundToNDecimals(scaledVal, varElement.getNumberDecimalPlaces()));
-            }
-
-            // Final Variant Construction and Validation
-            ArrayList<Element> finalVariantElements = new ArrayList<>();
-            double finalSum = 0.0;
-            boolean constraintViolatedAfterNormalization = false;
-
-            int varElemIdx = 0; // To iterate through tempVariableElements
-            for(Element rawElementFromFirstPass : currentRawVariantElements) {
-                Double minComp = rawElementFromFirstPass.getMin();
-                Double maxComp = rawElementFromFirstPass.getMax();
-                boolean isFixed = minComp != null && maxComp != null && minComp.doubleValue() == maxComp.doubleValue();
-
-                if (isFixed) {
-                    finalVariantElements.add(rawElementFromFirstPass); // Already a copy with fixed value
-                    finalSum += rawElementFromFirstPass.getPercentageComposition();
-                } else {
-                    // This must be one of the elements from tempVariableElements
-                    if (varElemIdx < tempVariableElements.size()) {
-                        Element processedVarElement = tempVariableElements.get(varElemIdx);
-                        finalVariantElements.add(processedVarElement);
-                        finalSum += processedVarElement.getPercentageComposition();
-
-                        // Final min/max check for this variable element
-                        if (processedVarElement.getMin() != null &&
-                                processedVarElement.getPercentageComposition() < (processedVarElement.getMin() - POST_NORM_CHECK_DELTA)) {
-                            logger.info("Sample discarded post-norm: Element " + processedVarElement.getSymbol() +
-                                    " (" + processedVarElement.getPercentageComposition() + ") below min " + processedVarElement.getMin());
-                            constraintViolatedAfterNormalization = true;
-                            break;
-                        }
-                        if (processedVarElement.getMax() != null &&
-                                processedVarElement.getPercentageComposition() > (processedVarElement.getMax() + POST_NORM_CHECK_DELTA)) {
-                            logger.info("Sample discarded post-norm: Element " + processedVarElement.getSymbol() +
-                                    " (" + processedVarElement.getPercentageComposition() + ") above max " + processedVarElement.getMax());
-                            constraintViolatedAfterNormalization = true;
-                            break;
-                        }
-                        varElemIdx++;
-                    } else {
-                        // Should not happen if logic is correct
-                        logger.severe("Mismatch between currentRawVariantElements and tempVariableElements. Discarding sample.");
-                        constraintViolatedAfterNormalization = true;
-                        break;
-                    }
-                }
-            }
-
-            if (constraintViolatedAfterNormalization) {
-                continue;
-            }
-
-            finalSum = CommonUtils.roundToNDecimals(finalSum, 2); // Round final sum before check
-            if (Math.abs(finalSum - 100.0) > FINAL_SUM_TOLERANCE) {
-                logger.info("Final sum " + finalSum + " is too far from 100. Discarding sample: " + commonUtils.buildCompositionString(finalVariantElements));
-                continue;
-            }
-
-            // If all checks pass
-            logger.info("New composition added: " + commonUtils.buildCompositionString(finalVariantElements));
-            variations.add(finalVariantElements);
-        }
-    }
-
+    @Deprecated
     public void getUniformDistribution(int index, List<Element> original, double varyBy, double limit,
                                        double currentSum, List<Element> currentCombo,
                                        List<List<Element>> results) {
