+++ src/main/java/com/medals/libsdatagenerator/service/LIBSDataService.java
@@ -198,79 +198,6 @@ public class LIBSDataService {
         return elementsList;
     }
 
-    public List<List<Element>> generateCompositionalVariations(List<Element> originalComposition,
-            double varyBy, double maxDelta, int variationMode,
-            int samples, String overviewGuid) {
-
-        List<List<Element>> compositions = new ArrayList<>();
-        if (originalComposition == null || originalComposition.isEmpty()) {
-            logger.warning("Original composition is null or empty. Cannot generate variations.");
-            return compositions; // Return empty list, or perhaps add the (empty) originalComposition if that's desired.
-        }
-        compositions.add(originalComposition); // Adding the original composition
-
-        boolean allElementsAreFixed = true;
-        for (Element el : originalComposition) {
-            Double minComp = el.getMin();
-            Double maxComp = el.getMax();
-            if (!(minComp != null && maxComp != null && minComp.equals(maxComp))) {
-                allElementsAreFixed = false;
-                break;
-            }
-        }
-
-        List<Element> effectiveComposition = originalComposition;
-        if (allElementsAreFixed) {
-            // originalComposition is not empty here due to the check at the beginning.
-            logger.info("All elements in the input composition are fixed. Applying fallback variation logic, ignoring X:X constraints for sampling.");
-            effectiveComposition = new ArrayList<>();
-            for (Element el : originalComposition) {
-                effectiveComposition.add(new Element(
-                    el.getName(),
-                    el.getSymbol(),
-                    el.getPercentageComposition(),
-                    null, // Effectively remove min/max for variation generation
-                    null,
-                    el.getAverageComposition()
-                ));
-            }
-        }
-
-        // Generate all combinations by Uniform distribution
-        // System.out.println("\nGenerating different combinations for the input composition (refer log for list)...");
-        logger.info("\nGenerating different combinations for the input composition (refer log for list)...");
-
-
-        if (variationMode == LIBSDataGenConstants.STAT_VAR_MODE_GAUSSIAN_DIST) {
-            CompositionalVariations.getInstance().gaussianSampling(
-                effectiveComposition,
-                maxDelta,
-                samples,
-                compositions);
-
-        } else if (variationMode == LIBSDataGenConstants.STAT_VAR_MODE_DIRICHLET_DIST) {
-            CompositionalVariations.getInstance().dirichletSampling(
-                effectiveComposition,
-                overviewGuid,
-                samples,
-                compositions
-            );
-
-        } else { // For uniform distribution
-            // Start backtracking with an empty "current combo" and a running sum of 0
-            CompositionalVariations.getInstance().getUniformDistribution(
-                0,
-                effectiveComposition,
-                varyBy,
-                maxDelta, // 'limit'
-                0.0,
-                new ArrayList<Element>(),
-                compositions);
-        }
-
-        return compositions;
-    }
-
     public void generateDataset(List<List<Element>> compositions, String minWavelength, String maxWavelength,
             String savePath, boolean appendMode, boolean forceFetch) {
